# Replace copier.Copy with optimized manual deep copy in URL.Clone (#2678)