@Nexusrex18 2025-03-18
@No-SilverBullet 
Rebased `issue-2678` onto `upstream/develop` (88823f4a) for CI sync. Benchmarks: 303 µs/op (avg of 5 runs), ~1.41x faster than `copier.Copy`. All `common/` tests pass locally (Windows & Ubuntu WSL, including `-race`), but Ubuntu CI fails: `make: *** [Makefile:59: test] Error 1`. This was fixed in #2795  by `develop`, but persists here.